extract starting debug server from language server
